refactor(backend): extract test set crud module
Finishes the crud split. The TestSet functions move into crud/test_set.py
and crud/__init__.py is left empty, so callers switch to direct submodule
imports.

Emptying __init__.py also removes the names it re-exported by accident,
which several modules were relying on: crud.get_item_detail, crud.include
and crud.delete_item are crud_utils/query_utils names, and
crud.schemas.TestConfigurationUpdate reached schemas through the package.
Each now imports from its real home.

Also restores delete_test_configuration, which went missing when
test_configuration was extracted -- DELETE /test_configurations/{id} has
been raising AttributeError since.

test_service_security.py passed the crud package itself into its table of
functions to check for an organization_id parameter. The loop guards with
hasattr, so an empty __init__.py would not fail it -- the five test set
assertions would just quietly stop checking anything. Points them at
test_set_crud.
