+++ apps/backend/src/rhesis/backend/app/crud/explorer.py
@@ -1,9 +1,5 @@
 """CRUD operations specific to Explorer test sets.
 
-Part of the incremental split of the ``crud`` monolith: ``crud/__init__.py`` still holds
-the bulk of the functions, and per-entity modules like this one take over as the code
-around them is touched.
-
 Every function here flushes and never commits -- the request session owns the commit
 (see ``get_db_with_tenant_variables`` in ``database.py``).
 """
@@ -17,13 +13,11 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session, contains_eager, joinedload
 
 from rhesis.backend.app import models, schemas
-
-# _TEST_SET_RELATED_FIELDS is imported rather than relocated: three other functions in
-# the monolith use the same tuple, and moving it would mean deciding a new home for a
-# TestSet-wide constant while this module only owns the Explorer slice.
-# crud/__init__.py never imports this module, so the parent-package import is cycle-free.
-from rhesis.backend.app.crud import _TEST_SET_RELATED_FIELDS
 from rhesis.backend.app.crud.embedding import create_embedding, get_embedding_by_hash
+
+# _TEST_SET_RELATED_FIELDS lives with the test-set CRUD: the test-set reads all share the
+# tuple, and this module only owns the Explorer slice.
+from rhesis.backend.app.crud.test_set import _TEST_SET_RELATED_FIELDS
 from rhesis.backend.app.models.enums import ModelType
 from rhesis.backend.app.models.test import test_test_set_association
 from rhesis.backend.app.schemas.explorer_metadata import (
