+++ apps/backend/src/rhesis/backend/jobs/execution/modes.py
@@ -9,7 +9,7 @@ import logging
 
 from sqlalchemy.orm import Session
 
-from rhesis.backend.app import crud
+from rhesis.backend.app import schemas
 from rhesis.backend.app.crud import test_configuration as test_configuration_crud
 from rhesis.backend.app.models.test import Test
 from rhesis.backend.app.models.test_configuration import TestConfiguration
@@ -90,7 +90,7 @@ def set_execution_mode(
         test_configuration_crud.update_test_configuration(
             db,
             test_config.id,
-            crud.schemas.TestConfigurationUpdate(**update_data),
+            schemas.TestConfigurationUpdate(**update_data),
             organization_id=organization_id,
             user_id=user_id,
         )
