+++ apps/backend/src/rhesis/backend/jobs/test_set.py
@@ -1,8 +1,8 @@
 import logging
 from typing import Any, List, Optional, Union
 
-from rhesis.backend.app import crud
 from rhesis.backend.app.constants import TestSetType
+from rhesis.backend.app.crud import test_set as test_set_crud
 from rhesis.backend.app.crud import user as user_crud
 from rhesis.backend.app.database import get_db_with_tenant_variables
 from rhesis.backend.app.models.enums import NotificationEventType
@@ -65,7 +65,7 @@ def count_test_sets(self):
         # Use tenant-aware database session with explicit organization_id and user_id
         with get_db_with_tenant_variables(org_id or "", user_id or "", project_id or "") as db:
             # Get all test sets with the proper tenant context
-            test_sets = crud.get_test_sets(db, organization_id=org_id, user_id=user_id)
+            test_sets = test_set_crud.get_test_sets(db, organization_id=org_id, user_id=user_id)
             total_count = len(test_sets)
             self.log_with_context("info", "Total test sets counted", total_count=total_count)
