+++ tests/backend/crud/test_test_set_crud.py
@@ -11,8 +11,9 @@ Run with: python -m pytest tests/backend/crud/test_test_set_crud.py -v
 import pytest
 from sqlalchemy.orm import Session
 
-from rhesis.backend.app import crud, models
+from rhesis.backend.app import models
 from rhesis.backend.app.crud import test as test_crud
+from rhesis.backend.app.crud import test_set as test_set_crud
 from rhesis.backend.app.utils.database_exceptions import ItemDeletedException
 
 
@@ -34,12 +35,12 @@ class TestTestSetSoftDeleteContract:
         test_db.refresh(test_set)
         test_set_id = test_set.id
 
-        crud.delete_test_set(
+        test_set_crud.delete_test_set(
             test_db, test_set_id, organization_id=test_org_id, user_id=authenticated_user_id
         )
 
         with pytest.raises(ItemDeletedException):
-            crud.get_test_set(test_db, test_set_id, organization_id=test_org_id)
+            test_set_crud.get_test_set(test_db, test_set_id, organization_id=test_org_id)
 
     def test_get_test_set_by_nano_id_or_slug_raises_for_deleted(
         self, test_db: Session, test_org_id: str, authenticated_user_id: str
@@ -56,17 +57,19 @@ class TestTestSetSoftDeleteContract:
         test_set_id = test_set.id
         slug = test_set.slug
 
-        crud.delete_test_set(
+        test_set_crud.delete_test_set(
             test_db, test_set_id, organization_id=test_org_id, user_id=authenticated_user_id
         )
 
         with pytest.raises(ItemDeletedException):
-            crud.get_test_set_by_nano_id_or_slug(test_db, slug, organization_id=test_org_id)
+            test_set_crud.get_test_set_by_nano_id_or_slug(
+                test_db, slug, organization_id=test_org_id
+            )
 
     def test_get_test_set_returns_none_for_nonexistent(self, test_db: Session, test_org_id: str):
         import uuid
 
-        result = crud.get_test_set(test_db, uuid.uuid4(), organization_id=test_org_id)
+        result = test_set_crud.get_test_set(test_db, uuid.uuid4(), organization_id=test_org_id)
         assert result is None
 
 
@@ -75,7 +78,7 @@ class TestTestSetSoftDeleteContract:
 class TestUpdateTestSetAttributesSoftDeleteHandling:
     """update_test_set_attributes must still no-op when a linked test set is deleted.
 
-    Regression test: crud.get_test_set now raises ItemDeletedException instead of
+    Regression test: test_set_crud.get_test_set now raises ItemDeletedException instead of
     returning None, so update_test_set_attributes (called by test_crud.update_test for
     every test set a test belongs to) must catch that itself -- otherwise updating
     a test would fail with a 410 just because an unrelated linked test set was
@@ -123,7 +126,7 @@ class TestUpdateTestSetAttributesSoftDeleteHandling:
         )
         test_db.commit()
 
-        crud.delete_test_set(
+        test_set_crud.delete_test_set(
             test_db, test_set.id, organization_id=test_org_id, user_id=authenticated_user_id
         )
