+++ tests/backend/test_secret_equality.py
@@ -67,8 +67,8 @@ SENSITIVE_MARKERS = (
 ALLOWED_SITES: frozenset[tuple[str, int]] = frozenset(
     {
         # content_hash is a SHA-based fingerprint for version dedup, not a secret
-        ("apps/backend/src/rhesis/backend/app/services/experiment.py", 150),
-        ("apps/backend/src/rhesis/backend/app/services/experiment.py", 206),
+        ("apps/backend/src/rhesis/backend/app/services/experiment.py", 152),
+        ("apps/backend/src/rhesis/backend/app/services/experiment.py", 208),
         # auth_token_project_id is a UUID project reference, not a secret token;
         # comparing it to another project UUID is safe (no timing oracle risk)
         ("apps/backend/src/rhesis/backend/app/services/connector/manager.py", 1005),
